refactor: apply .enum suffix to enums folder and modify leftover relative paths, refs AC-1266
