+++ libs/common/src/enums/device-type.enum.ts
