+++ libs/common/src/enums/encryption-type.enum.ts
