+++ libs/common/src/enums/hash-purpose.enum.ts
