+++ libs/common/src/enums/html-storage-location.enum.ts
