+++ libs/common/src/enums/index.ts
@@ -1,23 +1,23 @@
-export * from "./client-type";
-export * from "./device-type";
-export * from "./encrypted-export-type";
-export * from "./encryption-type";
-export * from "./event-system-user";
-export * from "./event-type";
-export * from "./field-type";
-export * from "./file-upload-type";
-export * from "./hash-purpose";
-export * from "./html-storage-location";
+export * from "./client-type.enum";
+export * from "./device-type.enum";
+export * from "./encrypted-export-type.enum";
+export * from "./encryption-type.enum";
+export * from "./event-system-user.enum";
+export * from "./event-type.enum";
+export * from "./field-type.enum";
+export * from "./file-upload-type.enum";
+export * from "./hash-purpose.enum";
+export * from "./html-storage-location.enum";
 export * from "./http-status-code.enum";
-export * from "./kdf-type";
-export * from "./key-suffix-options";
-export * from "./linked-id-type";
-export * from "./log-level-type";
-export * from "./native-messaging-version";
-export * from "./notification-type";
-export * from "./product-type";
-export * from "./secure-note-type";
-export * from "./state-version";
-export * from "./storage-location";
-export * from "./theme-type";
-export * from "./uri-match-type";
+export * from "./kdf-type.enum";
+export * from "./key-suffix-options.enum";
+export * from "./linked-id-type.enum";
+export * from "./log-level-type.enum";
+export * from "./native-messaging-version.enum";
+export * from "./notification-type.enum";
+export * from "./product-type.enum";
+export * from "./secure-note-type.enum";
+export * from "./state-version.enum";
+export * from "./storage-location.enum";
+export * from "./theme-type.enum";
+export * from "./uri-match-type.enum";
