+++ libs/common/src/enums/key-suffix-options.enum.ts
