+++ libs/common/src/enums/native-messaging-version.enum.ts
