+++ libs/common/src/enums/product-type.enum.ts
