+++ libs/common/src/enums/secure-note-type.enum.ts
