+++ libs/common/src/enums/storage-location.enum.ts
