+++ libs/common/src/vault/models/view/identity.view.ts
@@ -1,6 +1,6 @@
 import { Jsonify } from "type-fest";
 
-import { IdentityLinkedId as LinkedId } from "../../../enums/linked-id-type";
+import { IdentityLinkedId as LinkedId } from "../../../enums";
 import { linkedFieldOption } from "../../../misc/linkedFieldOption.decorator";
 import { Utils } from "../../../misc/utils";
